ASOC, ivm6303, fix: save Vs/Is enable state before turning speaker on
The start speaker sequence can switch vs/is enable off to avoid pops.
We must save the Vs/Is enable status before running the sequence and
then restore it. Note that we also always enable Is too. In case of
capture only (i.e. before enabling the speaker), only Vs = 1 makes sense,
because Is does not work. After playback start, Is can be turned on too.